+++ app/src/main/java/net/elenx/fishmash/updaters/WordsUpdater.java
@@ -67,6 +67,11 @@ public class WordsUpdater extends FishmashUpdater
     @Override
     public void save()
     {
+        if(words.size() < 1)
+        {
+            return;
+        }
+
         WordsDAO wordsDAO = new WordsDAO(optionsActivity);
         wordsDAO.truncate();
         wordsDAO.insert(words);
